Adiciona uso de manipulação de requisições por usuário

+++ musicas/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponseNotFound
 from .models import Gravadora
 
 def home(request):
@@ -9,6 +10,9 @@ def list_records(request):
     return render(request, 'list_gravadoras.html', {"gravadoras": gravadoras})
 
 def create_records(request):
+    if not request.user.is_authenticated:
+        # return HttpResponseNotFound("")
+        return redirect('lista-gravadoras')
     if request.method=="POST":
         nome = request.POST.get('nome')
         endereco = request.POST.get('endereco')
@@ -21,6 +25,9 @@ def create_records(request):
 
 
 def update_records(request, id):
+    if not request.user.is_authenticated:
+        # return HttpResponseNotFound("")
+        return redirect('lista-gravadoras')
     gravadora = get_object_or_404(Gravadora, pk=id)
     if request.method=="POST":
         nome = request.POST.get('nome')
@@ -33,6 +40,9 @@ def update_records(request, id):
 
 
 def remove_records(request, id):
+    if not request.user.is_authenticated:
+        # return HttpResponseNotFound("")
+        return redirect('lista-gravadoras')
     gravadora = get_object_or_404(Gravadora, pk=id)
     gravadora.delete()
     return redirect('lista-gravadoras')
